Feature/dispatcher

+++ src/main/resources/src/components/MainPage/Requests/TableView/TableView.jsx
@@ -19,15 +19,21 @@ const TableView = (props) => {
     }
 
     const searchQuery = (data) => {
-        // return data.filter(item => item.products[0].name.toLowerCase().includes(props.searchQuery.toLowerCase()))
-        return data.filter(item => item.responsible.toLowerCase().includes(props.searchQuery.toLowerCase()))
-        // return data.filter(item => {
-        //     return (
-        //         item.requestProducts.length !== 0 && item.requestProducts[0].product.name !== null
-        //             ? item.requestProducts[0].product.name.toLowerCase().includes(props.searchQuery.toLowerCase())
-        //             : item.status.toLowerCase().includes(props.searchQuery.toLowerCase())
-        //     )
-        // })
+        const query = props.searchQuery.toLowerCase();
+        return data.filter(item => {
+            return (
+                (item.requestProducts.length !== 0 && item.requestProducts[0].name !== null)
+                    ? (
+                        item.requestProducts[0].name.toLowerCase().includes(query) ||
+                        item.id.toString().includes(query) ||
+                        formatDateString(item.date).includes(query) ||
+                        item.codeWord.toLowerCase().includes(query) ||
+                        item.status.toLowerCase().includes(query) ||
+                        item.responsible.toLowerCase().includes(query)
+                    )
+                    : item.status.toLowerCase().includes(query)
+            )
+        })
     }
 
     const handleStatusChange = (event) => {
@@ -143,10 +149,16 @@ const TableView = (props) => {
             </div>
             {sortRequests(props.data).map((request, request_id) => (
                 <div key={request_id} className={"tableview_requests__row " +
-                    (request.status === "Не готово" && "tableview_requests__row--status_not_ready" ||
-                        request.status === "В процессе" && "tableview_requests__row--status_in_progress" ||
-                        request.status === "Готово к отгрузке" && "tableview_requests__row--status_ready" ||
-                        request.status === "Отгружено" && "tableview_requests__row--status_shipped")
+                    (
+                        request.status === "Проблема" && "tableview_requests__row--status_problem" ||
+                        request.status === "Материалы" && "tableview_requests__row--status_materials" ||
+                        request.status === "Ожидание" && "tableview_requests__row--status_waiting" ||
+                        request.status === "В производстве" && "tableview_requests__row--status_in_production" ||
+                        request.status === "Готово" && "tableview_requests__row--status_ready" ||
+                        request.status === "Отгружено" && "tableview_requests__row--status_shipped" ||
+                        request.status === "Приоритет" && "tableview_requests__row--status_priority" ||
+                        request.status === "Завершено" && "tableview_requests__row--status_completed"
+                    )
                 }>
                     <div className="tableview_requests__col">{request.id}</div>
                     <div className="tableview_requests__col">{formatDateString(request.date)}</div>
@@ -154,9 +166,9 @@ const TableView = (props) => {
                         {request.requestProducts.map((item, index) => {
                             return (
                                 <div className="tableview_requests__sub_row" style={{ height: `calc(${100 / request.requestProducts.length}%)` }}>
-                                    <div className="tableview_requests__sub_col">{item.product && item.product.name}</div>
+                                    <div className="tableview_requests__sub_col">{item.name}</div>
                                     <div className="tableview_requests__sub_col">{item.packaging}</div>
-                                    <div className="tableview_requests__sub_col">{item.quantity + " " + (item.product && item.product.unit)}</div>
+                                    <div className="tableview_requests__sub_col">{item.quantity}</div>
                                 </div>
                             )
                         })}
@@ -179,16 +191,20 @@ const TableView = (props) => {
                             defaultValue={request.status}
                             onChange={handleStatusChange}
                         >
-                            <option>Не готово</option>
-                            <option>В процессе</option>
-                            <option>Готово к отгрузке</option>
-                            <option>Отгружено</option>
+                            <option value="Проблема">Проблема</option>
+                            <option value="Материалы">Материалы</option>
+                            <option value="Ожидание">Ожидание</option>
+                            <option value="В производстве">В производстве</option>
+                            <option value="Готово">Готово</option>
+                            <option value="Завершено">Завершено</option>
+                            <option value="Отгружено">Отгружено</option>
+                            <option value="Приоритет">Приоритет</option>
                         </select>
                     </div>
                     <div className="tableview_requests__actions">
                         <Link to={"/requests/view/" + request.id} className="tableview_requests__action" >Просмотр</Link>
                         {props.userHasAccess(['ROLE_ADMIN', 'ROLE_MANAGER']) && <Link to={"/requests/edit/" + request.id} className="tableview_requests__action">Редактировать</Link>}
-                        {/* <div data-id={request.id} className="tableview_requests__action" onClick={props.deleteItem}>Удалить</div> */}
+                        {props.userHasAccess(['ROLE_ADMIN']) && <div data-id={request.id} className="tableview_requests__action" onClick={props.deleteItem}>Удалить</div>}
                     </div>
                 </div>
             ))}
